reorganizing directories and fixing a tests
