+++ apps/braze/src/components/generate/CodeBlocksStep.tsx
@@ -1,11 +1,11 @@
 import { useEffect, useState } from 'react';
-import WizardFooter from './WizardFooter';
+import WizardFooter from '../WizardFooter';
 import { Box, Button, Paragraph, Skeleton, Subheading } from '@contentful/f36-components';
-import Splitter from './Splitter';
+import Splitter from '../Splitter';
 import tokens from '@contentful/f36-tokens';
-import CodeBlock from './CodeBlock';
-import { Entry } from '../fields/Entry';
-import OctagonValidationMessage from './OctagonValidationMessage';
+import CodeBlock from '../CodeBlock';
+import { Entry } from '../../fields/Entry';
+import OctagonValidationMessage from '../OctagonValidationMessage';
 
 type CodeBlocksStepProps = {
   entry: Entry;
