+++ apps/braze/test/mocks/mockSdk.ts
@@ -17,7 +17,7 @@ const mockSdk: any = {
   parameters: {
     installation: {
       apiKey: 'test-apiKey',
-      contentfulApiKey: 'test-contentful-apiKey',
+      contentfulApiKey: 'test-apiKey',
       brazeApiKey: 'test-braze-apiKey',
     },
     invocation: {
